Added rankup notifiactions for members when they come online
The message is sent to all members who are eligible for their rankup test every week when they come online. The week-delay betwen pesters does not scale with rank; will not send messages to members whose last rankup is unknown. A log message is sent to the #botspam channel on Spawner Swarm and the Guild Master role is tagged (with each Pester).

Also fixed a stupid in the guildMemberAdd event.